Former-commit-id: 6e92bf62bca6f7c3d241575845d60a5746c96569

+++ applications/tests/test_image/test_image_main.cpp
@@ -316,7 +316,7 @@ TEST_F( ImageTest, movePointerTo)
     {
         img2.read(auxFn, DATA, n);
 
-        for (int k = 1; k <= aDim.zdim; ++k)
+        for (size_t k = 1; k <= aDim.zdim; ++k)
         {
             img1.movePointerTo(k, n);
             img2.movePointerTo(k);
